mlx5: Add DCS caps support
DCS is an offload to SW load balancing of DC initiator work requests.

A single DCI can be connected to only one target at the time and can't
start a new connection until the previous work request is completed.
This limitation will cause a delay when the initiator process needs to
transfer data to multiple targets at the same time.
The SW solution is to use a process that handling and spreading the work
request on many DCIs according to destinations.

This feature is an offload to this process and coming to reduce the load
from the CPU and improve the performance

Signed-off-by: Lior Nahmanson <liorna@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>