pyverbs: Added DCS modify QP support
In case a CQE with error is generated for some WQE, the dci stream
channel that attached to this WQE is from now in an error state and all
the subsequent WQEs that have the same dci stream channel id need to be
flushed also.

SW can reset dci stream channel state using RTS2RTS_QP command so the
dci stream channel will no longer be in error state.

Signed-off-by: Lior Nahmanson <liorna@nvidia.com>
Signed-off-by: Edward Srouji <edwards@nvidia.com>